compile i386 backend needed by libebl
Libcpu also needed to be pulled in.

The generated files are not suffixed with rh because ebl opens them
dynamically, so we would need to change the sources.

Needed to add libcpu/i386_dis.h because it can't be generated by
android build system (can't compile a file that generates
the header).

For: AXIA-972
Depends-Change-Id: Ib60887f0c245b64247b77101cc1870eb0087f82e
Change-Id: I8aae09ce7faec71172fbfe115a13627a588d29fe
Signed-off-by: Dragos Tatulea <dragos.tatulea@intel.com>